+++ library/src/main/java/library/CardDrawSpaces.java
@@ -1,0 +1,5 @@
+package library;
+//this class handles community chest cards and chance card spaces, using json to load cards and apply a random one to playerx.
+
+public class CardDrawSpaces {
+}