+++ library/src/main/java/library/GoSpace.java
@@ -1,0 +1,11 @@
+package library;
+
+/**
+ * this class handles GO Space
+ * As Player passes go add 200 to budget.
+ * Can possibly add to space superclass.
+ */
+public class GoSpace {
+
+    //more to come
+}