+++ library/src/main/java/library/UtilitySpaces.java
@@ -1,0 +1,8 @@
+package library;
+/*
+*Class to Manage Utility spaces, They have special rent rules and include electric company and Water Works
+*
+*
+ */
+public class UtilitySpaces extends Space{
+}