+++ library/src/test/java/libraryTest/DiceTest.java
@@ -1,0 +1,22 @@
+package libraryTest;
+
+import library.Dice;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.Assertions;
+
+
+
+public class DiceTest {
+    Dice dice =new Dice();
+
+    @Test
+    public void testroll(){
+        Dice dice =new Dice();
+        int F1 = dice.getFaceValue1();
+        int F2 = dice.getFaceValue2();
+        int calculatedSum= F1+F2;
+        Assertions.assertEquals(calculatedSum,dice.getDiceTotal());
+    }
+
+
+}