+++ library/src/test/java/libraryTest/GameStateTest.java
@@ -1,0 +1,4 @@
+package libraryTest;
+
+public class GameStateTest {
+}