+++ server/src/main/java/Server/ServerInit.java
@@ -1,0 +1,37 @@
+package Server;
+
+import io.javalin.Javalin;
+import org.eclipse.jetty.server.Server;
+import org.eclipse.jetty.util.thread.QueuedThreadPool;
+import library.Player;
+
+import static io.javalin.apibuilder.ApiBuilder.get;
+
+/**
+ * This class initializes the server and displays result message.
+ */
+
+public class ServerInit {
+    public static void main(String[] args) {
+        QueuedThreadPool queuedThreadPool =
+                new QueuedThreadPool(200, 8, 60000);
+
+        Javalin app = Javalin.create(config ->
+                config.server(() ->
+                        new Server(queuedThreadPool))).start(7000);
+
+        app.get("/", ctx -> ctx.result("Server Launch Successful"));
+        app.routes(() -> {
+
+            //get("/api/state", ctx -> ctx.json(Player.ReadGameState()));
+            post("/api/update", ctx -> {
+
+            });
+
+
+            get("/api/status", ctx -> {
+                ctx.result("OK");
+            });
+        });
+    }
+}