+++ server/src/test/java/serverTest/OneEqualsOne.java
@@ -1,0 +1,12 @@
+package serverTest;
+
+
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.Assertions.*;
+
+public class OneEqualsOne {
+    @Test public void constructorTest() {
+        Assertions.assertEquals(1, 1);
+    }
+}