+++ CHANGELOG.md
@@ -4,6 +4,46 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [Unreleased]
+
+
+### Added
+-Images for Each Spaces
+
+-More Tests
+
+-Gson Processing 
+
+-Get and Post Sections
+
+-gitignore file added
+
+###Changed
+
+-Updated Json
+
+
+
+
+## [2.0.0]
+
+### Added
+-Game Board in JavaFX
+
+-Added Classes into Library
+
+### Changed
+-Connection Capabilites further developed
+
+-Login submission was converted to FXML
+
+-Updated Dice Class
+
+-FXML implemented
+
+-Post & Get Methods Implemented
+
 ## [Unreleased]
 
 ## [0.0.1] - 2021-9-30
@@ -15,17 +55,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-## [Unreleased]
+## [Released]
 
 ## [1.0.0] - 2021-9-30
 
 This will be the work we do from the first sprint.
 ### Added
-	-instert new feature descriptions here
+	-server and Client java files added
+    -Module info added
+    -GUI for entering address and port made
 
 
-### Changeed
-	-put descriptions of feature descriptions here
 
+### Changed
+    -Server initialied and runs
+    -client initialized and runs
+    -GUI created and works
 ### Removed
 	-list what you remove here